+++ src/meltano/cli/discovery.py
@@ -15,7 +15,7 @@ from . import cli
 from .params import pass_project
 
 discoverable_plugin_types = [
-    ptype for ptype in list(PluginType) if ptype is not PluginType.CLIS
+    ptype for ptype in list(PluginType) if ptype is not PluginType.UTILITIES
 ]